refactor build-in resolvers
