+++ omegaconf/resolvers/__init__.py
@@ -1,0 +1,5 @@
+from .env import legacy_env
+
+__all__ = [
+    "legacy_env",
+]
