+++ omegaconf/resolvers/oc/__init__.py
@@ -1,0 +1,7 @@
+from ._decode import decode
+from ._env import env
+
+__all__ = [
+    "env",
+    "decode",
+]
