+++ setup.py
@@ -54,6 +54,9 @@ with open("README.md", "r") as fh:
             "omegaconf",
             "omegaconf.grammar",
             "omegaconf.grammar.gen",
+            "omegaconf.resolvers",
+            "omegaconf.resolvers.oc",
+            "omegaconf.resolvers.oc.dict",
             "pydevd_plugins",
             "pydevd_plugins.extensions",
         ],
